+++ src/analytic/penning_trap_uniform.jl
@@ -13,7 +13,7 @@ B (x,y,z) = B_0 \, ( 0, 0, 1)^T ,
 ```
 and the electrostatic potential given by
 ```math
-\varphi (x,y,z) = E_0 \, ( x^2 / 2 + y^2 / 2 - z^2) ,
+\varphi (x,y,z) = - E_0 \, ( x^2 / 2 + y^2 / 2 - z^2) ,
 ```
 resulting in the electric field with covariant components
 ```math
@@ -25,8 +25,6 @@ Parameters:
 * `E₀`: E-field strength
 """
 module PenningTrapUniform
-
-    using RecipesBase
 
     import ..ElectromagneticFields
     import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments