+++ test/runtests.jl
@@ -3,3 +3,4 @@ using ElectromagneticFields
 using Test
 
 include("test_analytic.jl")
+include("test_plots.jl")